add --auto: quote ssh-keygen filename for windows
Windows home directories basically always have spaces in them. So we
should make sure we're expecting that. I put quotes around the filename
in the `ssh-keygen` command, because without them a Windows client was
unable to automatically generate an ssh key.

If you don't need to support Ruby 1.8.7, Ruby's `.shellescape` would be
a more robust solution.